add web socket connection, fix animation position
