+++ client/components/Animation/Animation.jsx
@@ -10,12 +10,12 @@ import MenuPanel from '../MenuPanel/menuPanel';
 const Animation = () => {
   const { data, setData } = useContext(BreathContext)
 
-  console.log('selectedOption', data.selectedOption);
+  // console.log('selectedOption', data.selectedOption);
   const style = data.selectedOption;
 
   return (
     <div>
-      <div className ="graphic">
+      <div className ="container">
         <div className={style}></div>
         <div className={style}></div>
         <div className={style}></div>
