+++ client/components/Animation/SquareAnimation.module.scss
@@ -1,11 +1,11 @@
 
-#container {
-  width: 600px;
-  height: 400px;
+.container {
+  width: 100%;
+  // height: 400px;
   display: flex;
   align-items: center;
   justify-content: center;
-  overflow: hidden;
+  // overflow: hidden;
   position: relative;
 }
