+++ client/context/Context.jsx
@@ -19,6 +19,7 @@ export const BreathContextProvider = (props) => {
     logged_in: false,
     f_name: '',
     ws_connected: false,
+    user_count: 0,
   });
 
   const [signup, setSignup] = useState({
